backend/requirements.txt
# Compatibility install path only.
# Backend dependency source of truth is `pyproject.toml` + `uv.lock`.

fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
psycopg2-binary>=2.9,<3
alembic>=1.13,<2
pydantic-settings>=2.2,<3
email-validator>=2.1,<3
pgvector>=0.3,<1

# Agent orchestration
langgraph>=0.2,<1
langchain-core>=0.3,<1
langchain-google-genai>=2,<3

# Auth
bcrypt>=4,<6
python-jose[cryptography]>=3.3,<4
slowapi>=0.1.9,<1

httpx>=0.27,<1
elevenlabs>=1,<2
google-genai>=1,<2
pytest>=8,<9
pytest-asyncio>=0.23,<1
ruff>=0.3,<1
google-generativeai>=0.1,<1
pdfplumber>=0.7,<1
